Various small bugfixes and tweaks...
Found usefull for running our internal cacofonisk based project.

Made the FileRunner() accept filenames directly...

Made the FileRunner() accept filenames directly as well as lists of
filenames. And do input checking.

Added declaration of reporter in example.

Added missing methods in BaseReporter

Methods that are already used in the ChannelManager

Moved the BaseTestCase and SilentReporter

I moved them to the utils folder. Inspired by how we do things in
VoIPGRID.

Added comment and help to tests

using import json

Makes things more readable when calling json.load()
