+++ tests/replaytest.py
@@ -1,12 +1,8 @@
 import os
-import sys
-
-from datetime import datetime
-from json import load
-from unittest import TestCase
 
 from cacofonisk.channel import CallerId, ChannelManager
 from cacofonisk.runners.file_runner import FileRunner
+from cacofonisk.utils.testcases import BaseTestCase, SilentReporter
 
 
 class MockChannelManager(ChannelManager):
@@ -42,30 +38,6 @@ class MockChannelManager(ChannelManager):
              'party1': party1, 'party2': party2})
 
 
-class SilentReporter(object):
-    def __init__(self):
-        self.silent = True
-
-    def on_event(self, event):
-        pass
-
-    def on_b_dial(self, caller, callee):
-        pass
-
-    def on_transfer(self, redirector, party1, party2):
-        pass
-
-    def trace_ami(self, event):
-        pass
-
-    def trace_msg(self, msg):
-        if not self.silent:
-            sys.stdout.write('{}: {}\n'.format(
-                datetime.now().strftime('%Y-%m-%d %H:%M:%S.%f'),
-                msg))
-            sys.stdout.flush()
-
-
 class BogoRunner(object):
     def __init__(self, events, reporter):
         self.events = events
@@ -80,18 +52,6 @@ class BogoRunner(object):
                 channelmgr.on_event(event)
 
         self.channel_managers.append(channelmgr)
-
-
-class BaseTestCase(TestCase):
-    def open_file(self, filename, *args, **kwargs):
-        path = os.path.dirname(__file__)
-        filename = os.path.join(path, filename)
-        return open(filename, *args, **kwargs)
-
-    def load_events_from_disk(self, filename):
-        with self.open_file(filename, 'r') as f:
-            events = load(f)
-        return events
 
 
 class ChannelEventsTestCase(BaseTestCase):
